moved schedule into its own file
